Improve bounded recovery readability
